👷 Update upload/download artifacts action to v4
Update upload-artifact and download-artifact actions to v4

Replace deprecated v3 with the latest v4 version for both
upload-artifact and download-artifact actions in the CI/CD
workflow. This update ensures the use of the most recent and
supported versions of these GitHub Actions, potentially
benefiting from performance improvements and bug fixes.